Flask RESTful

+++ resources/item.py
@@ -1,4 +1,11 @@
-from flask_jwt import jwt_required
+
+from flask_jwt_extended import (
+    jwt_required,
+    get_jwt_claims,
+    jwt_optional,
+    get_jwt_identity,
+    fresh_jwt_required
+)
 from flask_restful import Resource, reqparse
 
 
@@ -18,13 +25,14 @@ class Item(Resource):
                         required=True,
                         help="This field cannot be left blank!")
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {'item': "Item not found"}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {
@@ -39,8 +47,11 @@ class Item(Resource):
             return {'message': "An error occurred inserting the item"}, 500  # internal server error
 
         return item.json(), 201
-
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin access required.'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -60,7 +71,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {'items': [item['name'] for item in items],
+                'message': 'More data available if you log in'}, 200
 
  # return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
